requirements.txt
fastapi==0.115.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.61.1
greenlet==3.2.4
griffe==1.15.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
humanfriendly==10.0
idna==3.11
ImageIO
imagesize==1.4.1
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
modelscope==1.33.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
# nvidia-cublas-cu12==12.8.4.1
# nvidia-cuda-cupti-cu12==12.8.90
# nvidia-cuda-nvrtc-cu12==12.8.93
# nvidia-cuda-runtime-cu12==12.8.90
# nvidia-cudnn-cu12==9.10.2.21
# nvidia-cufft-cu12==11.3.3.83
# nvidia-curand-cu12==10.3.9.90
# nvidia-cusolver-cu12==11.7.3.90
# nvidia-cusparse-cu12==12.5.8.93
# nvidia-cusparselt-cu12==0.7.1
# nvidia-nccl-cu12==2.27.5
# nvidia-nvjitlink-cu12==12.8.93
# nvidia-nvshmem-cu12==3.3.20
# nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opencv-contrib-python
opencv-python
opencv-python-headless
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt-einsum==3.3.0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
#packaging==25.0
#paddleocr==3.3.2
#paddlepaddle==3.2.2
paddlex==3.3.13
pandas
parso
pexpect==4.9.0
pillow
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
posthog==5.4.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64
pyclipper
pycryptodome
pydantic
pydantic-settings
#pydantic_core
#Pygments
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
ruamel.yaml==0.19.1
safetensors==0.6.2
scikit-image==0.26.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.38.6
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1

ultralytics
ultralytics-thop
uritemplate
urllib3
uuid_utils
uvicorn
